+++ packages/backend/src/postgres.ts
@@ -67,7 +67,6 @@ import { MiUserKeypair } from '@/models/UserKeypair.js';
 import { MiUserList } from '@/models/UserList.js';
 import { MiUserListFavorite } from '@/models/UserListFavorite.js';
 import { MiUserListMembership } from '@/models/UserListMembership.js';
-import { MiUserMiniApp } from '@/models/UserMiniApp.js';
 import { MiUserNotePining } from '@/models/UserNotePining.js';
 import { MiUserPending } from '@/models/UserPending.js';
 import { MiUserProfile } from '@/models/UserProfile.js';
@@ -198,7 +197,6 @@ export const entities = [
 	MiUserList,
 	MiUserListFavorite,
 	MiUserListMembership,
-	MiUserMiniApp,
 	MiUserNotePining,
 	MiUserSecurityKey,
 	MiUsedUsername,